refactor: improve error handling and code consistency

+++ lib/features/todo/repositories/todo_repository_impl.dart
@@ -2,6 +2,8 @@ import 'package:to_do_app/features/todo/models/task_model/task_model.dart';
 import 'package:to_do_app/features/todo/repositories/todo_repository.dart';
 import 'package:to_do_app/features/todo/services/todo_service.dart';
 
+const int _kMaxRemoteTaskId = 200;
+
 class TodoRepositoryImpl implements TodoRepository {
   final TodoService _todoService;
 
@@ -21,13 +23,15 @@ class TodoRepositoryImpl implements TodoRepository {
   }
 
   @override
-  Future<TaskModel> updateTodo(int id, Map<String, dynamic> data) async {
+  Future<TaskModel?> updateTodo(int id, Map<String, dynamic> data) async {
+    if (id > _kMaxRemoteTaskId) return TaskModel.fromJson(data);
     final response = await _todoService.updateTodo(id, data);
     return TaskModel.fromJson(response);
   }
 
   @override
   Future<void> deleteTodo(int id) async {
+    if (id > _kMaxRemoteTaskId) return;
     await _todoService.deleteTodo(id);
   }
 }
